Add an initial transcoder implementation
Add a filesystem watcher that watches for events in the photos originals
dir and then pushes a job for each Create event.

The job is then picked up by a pool of workers and they call a system
executable (ffmpeg eventually) to process the video.

Still lots to work out, like listening for Remove events, and handling
events for changed files, or when events for a file that is already
being processed.