Process HDR videos with different arguments
Use ffprobe to determine if the video about to be transcoded is HDR or
not based on its color profile. We use different set of arguments to
pass to ffmpeg when transcoding hdr videos. By default it tries to do
tone mapping to preserve the color.

Transcoding HDR requires some additional filters that might not be
available in a default ffmpeg installation. Arguments can be changed
in the settings